requirements.txt
# Scientific AI Backplane - Python Dependencies
# Install with: pip install -r requirements.txt

# Core dependencies
celery>=5.3.0
redis>=4.6.0
requests>=2.31.0
pyyaml>=6.0
fastapi>=0.104.0
uvicorn>=0.24.0

# Scientific computing
numpy>=1.24.0
scipy>=1.11.0
pandas>=2.0.0
matplotlib>=3.7.0
h5py>=3.7.0

# Quantum chemistry
ase>=3.22.0
pyscf>=2.3.0

# Machine learning
torch>=2.0.0
mace-torch>=0.3.0

# File parsing (optional)
# pymatgen>=2023.0.0  # For advanced CIF parsing
